+++ packages/react-magma-dom/src/components/RadioGroup/index.tsx
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 
 import { useIsInverse } from '../../inverse';
 import { ThemeContext } from '../../theme/ThemeContext';
-import { omit, useGenerateId } from '../../utils';
+import { descriptionSuffix, omit, useGenerateId } from '../../utils';
 import { HiddenStyles } from '../../utils/UtilityStyles';
 import { InputMessage } from '../Input/InputMessage';
 import { Label } from '../Label';
@@ -130,7 +130,8 @@ export const RadioGroup = React.forwardRef<HTMLDivElement, RadioGroupProps>(
     } = props;
     const other = omit(['onBlur', 'onChange', 'onFocus', 'id'], rest);
 
-    const descriptionId = errorMessage || helperMessage ? `${id}__desc` : null;
+    const descriptionId =
+      errorMessage || helperMessage ? `${id}${descriptionSuffix}` : null;
 
     const theme = React.useContext(ThemeContext);