+++ packages/react-magma-dom/src/components/TimePicker/TimePicker.test.js
@@ -470,12 +470,11 @@ describe('TimePicker', () => {
 
     it('should use the i18n context defaults', () => {
       const { getByLabelText } = render(<TimePicker labelText="label" />);
-
       expect(
-        getByLabelText(`label, ${defaultI18n.timePicker.hoursAriaLabel}`)
+        getByLabelText(defaultI18n.timePicker.hoursAriaLabel)
       ).toBeInTheDocument();
       expect(
-        getByLabelText(`label, ${defaultI18n.timePicker.minutesAriaLabel}`)
+        getByLabelText(defaultI18n.timePicker.minutesAriaLabel)
       ).toBeInTheDocument();
     });
   });