+++ website/react-magma-docs/src/components/PageContent/index.js
@@ -226,13 +226,13 @@ export const PageContent = ({ children, componentName, type }) => {
         const designPatternsLink = designPatternDocs?.node.fields.slug;
         const dataVisualizationLink = dataVisualization?.node.fields.slug;
 
-        const hasDocs = !!(
+        const hasNavTabs = !!(
           apiDocs ||
           designDocs ||
           patternsDocs ||
-          designPatternDocs ||
-          dataVisualization
+          designPatternDocs
         );
+        const hasDocs = !!(hasNavTabs || dataVisualization);
 
         const apiNavTabToLink = patternsDocs
           ? patternsLink
@@ -283,30 +283,32 @@ export const PageContent = ({ children, componentName, type }) => {
             {hasDocs ? (
               <>
                 <StyledTabsContainer isInverse={isInverse}>
-                  <TabsWrapper>
-                    <StyledTabs aria-label="">
-                      {apiDocs || patternsDocs ? (
-                        <NavTab
-                          component={
-                            <Link to={apiNavTabToLink}>Implementation</Link>
-                          }
-                          isActive={type === NAV_TABS.API}
-                        />
-                      ) : (
-                        <></>
-                      )}
-                      {designDocs || designPatternDocs ? (
-                        <NavTab
-                          component={
-                            <Link to={designNavTabToLink}>Design</Link>
-                          }
-                          isActive={type === NAV_TABS.DESIGN}
-                        />
-                      ) : (
-                        <></>
-                      )}
-                    </StyledTabs>
-                  </TabsWrapper>
+                  {hasNavTabs && (
+                    <TabsWrapper>
+                      <StyledTabs aria-label="">
+                        {apiDocs || patternsDocs ? (
+                          <NavTab
+                            component={
+                              <Link to={apiNavTabToLink}>Implementation</Link>
+                            }
+                            isActive={type === NAV_TABS.API}
+                          />
+                        ) : (
+                          <></>
+                        )}
+                        {designDocs || designPatternDocs ? (
+                          <NavTab
+                            component={
+                              <Link to={designNavTabToLink}>Design</Link>
+                            }
+                            isActive={type === NAV_TABS.DESIGN}
+                          />
+                        ) : (
+                          <></>
+                        )}
+                      </StyledTabs>
+                    </TabsWrapper>
+                  )}
 
                   <StyledTabPanelsContainer>
                     <StyledTabPanel>